feat: refactor agents app
