+++ retail/agents/domains/agent_integration/exceptions.py
@@ -1,0 +1,14 @@
+from rest_framework.exceptions import APIException
+from rest_framework import status
+
+
+class GlobalRuleBadRequest(APIException):
+    status_code = status.HTTP_400_BAD_REQUEST
+
+
+class GlobalRuleUnprocessableEntity(APIException):
+    status_code = status.HTTP_422_UNPROCESSABLE_ENTITY
+
+
+class GlobalRuleInternalServerError(APIException):
+    status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
