+++ retail/agents/domains/agent_management/usecases/retrieve.py
@@ -1,0 +1,16 @@
+from django.db.models import QuerySet
+
+from uuid import UUID
+
+from rest_framework.exceptions import NotFound
+
+from retail.agents.domains.agent_management.models import Agent
+
+
+class RetrieveAgentUseCase:
+    @staticmethod
+    def execute(agent_uuid: UUID) -> QuerySet[Agent]:
+        try:
+            return Agent.objects.get(uuid=agent_uuid).prefetch_related("templates")
+        except Agent.DoesNotExist:
+            raise NotFound(f"Agent not found: {str(agent_uuid)}")
